Alliance list and Menu alliance function updated
Added USSR to the alliance list, stripped duplicate entries since length
is no concern now

Added nested alliance functionality, so an alliance can call and
alliance, ex: USSR is an alliance, and a member of the alliance Warsaw
Pact.

Moved the flatten function from magazine generator to Utilities, which
it technically is.