UI.GetItemCapacityForLauncher added, GetCompatibleItemList Filter pre-empted
added GetItemCpacityFor launcher to platform Interface and made callable
from python.

inserted a return ahead of the filtering for GetCompatibleItems list
until I am proficient enough to insert a check for year filtering in its
place or to perform the task differently and account for the whims of
the function calling it.