AddAlliances, GetAllianceRelations, Datum Ranges
Restored the datum range limitation for non missile weapons.  Only
missile weapons have the kinetic advantages of velocity and altitude in
enough quantity to significantly exceed the listed ranges.  So missiles
are not limited by this check, but other datum types are again limited.

Added a menu function to utilise the CreateAlliance function.  Modified
ToggleAlliance to also support the full 16 alliances.

Added GetAllianceRelationship to Scenario Interface to permit reading
the current relations.