db, icons, alliances update, reduced damage rand()
Many new class entries, the complete run of Russian and Chinese torpedo
boats, including all their exports, required torpedoes, and weapon
icons.

Added North Vietnam to the alliances.  Permits units to be assigned to
North Vietnam to still be considered when using Vietnam in 1975., after
their re-unification.

Refactored the randomness in the damage , its not 0% to 200% anymore,
rather, its 100%±30% now.  Might be high or low, but it should average
out the same as before.  Difference is that definitely fatal hits don't
have a chance to roll a 0.001 and be survivable now.  200% * 0.7 is
still 140%, your still gonna die even if you roll a zero.  142.8% damage
is the guaranteed fatal cut-off as a result, above that, chance cannot
roll low enough to save the unit.