Ver 1.28.0.3, sub/ship and dBedit fix/update
fixed the logic for accel/decel to better work with all possible
velocities and goals.

boosted deceleration with a crashstop doubling modifier

reverted to using BatteryCpacity_kJ for battery capacity since I cannot
find all references to it and the dBeditor was rather unhappy about not
finding it.

updated the sub tab in the database editor to correspond with the
database changes so that it can now work with the AIP update.